-- Reverted the ride count mechanic for individual rides on a visit. The Ride object is gone for good.
NOTE: I have recently decided to finish the project within this lifetime; therefore, I am willing to reduce the scope of the app and expect time savings of roughly around 60-90 years...
The idea behind the Ride object was to create a single object for every ride done on a visit - that would have allowed the user to track all sorts of stuff for every individual ride, like timestamp, ratings, queuetime, notes, specialities like rollback, evacuation, nightride, rides in the rain etc and so forth.
As it is theoretically possible to implement an option for the user to input number of trains, seats and rows per train for every coaster, the use of Ride objects would allow for tracking where exactly the user sat on the ride - in combination with the user's ratings, with enough data the app would have been able to suggest a riding position under the given circumstances. Usage of tags like 'airtime' or 'forceful' or the like would have been thinkable - but no: all that is gone (or nice to have - maybe I will come back to the idea when I have some more years to spare)...
As it is for now the users have to track that kind of stuff for themselfs by using the Notes field (yet to be implemented).