-- preparing introduction of OrphanElement CreditType (see notes below) -- AttractionCategories are now just named Categories -- added a bunch of contructors - makes mocking data easier
NOTE: At the time the app distinguishes between Coasters and 'other' Attractions - where Coasters are credits and all other attractions are not.
As this is a credit counter app I decided that this distinction is a little too basic and needs some love: introducing CreditTypes!
In near future Coaster, CustomCoaster and CoasterBlueprints will be removed from the app. Instead the user selects a CreditType when creating an Attraction. This allows the user to define and track all kinds of credit types, like DropTowers, Slides or BathroomFacilities and -of course- RollerCoasters...