-- further refactored GroupHeader handling in CRVA -- found out what IGroupHeader.groupElements are - yep, I am keeping them...
NOTE: I am not happy with the IGroupHeader interface - either I have to cast to GroupHeader in order to get the GroupElement, or SpecialGroupHeader has to implement getter returning null. Has a bad smell, that one... =( We will see what time will bring - the bright (hopefully near) future, when I will be a better developer. ^^