-- removed interface IOnSiteAttraction and class OnSiteAttraction -- renamed CustomAttractions to OnSiteAttractions -- removed interfaces IHasProperty, IHasManufacturer, IHasStatus, IHasCategory and IHasCreditType
NOTE: as with Blueprints gone StockAttractions are no more - and therefore no need for the IOnSiteAttraction interface: all Attractions that are not VisitedAttractions are OnSiteAttractions. So I removed the interface and renamed CustomAttractions to OnSiteAttractions. As all Attractions now have properties (was only Blueprint which hasn't) there is no need for the property interfaces any more.