-- introducing Models: Attraction can now hold a Model
Note:
Models basically replace IAttraction.StockAttractions. They are Elements with interface IProperty which hold a CreditType, a Category and a Manufacturer. When a Model is set on an Attraction, CreditType, Category and Manufacturer are set from Model in Attraction.
That way there is only one type of Attraction (apart from VisitedAttractions...) and all the properties are accessable as usual in (OnSite)Attractions. If a property is changed in the Model the reference should change so that the newly set property is referenced in the Attraction. That, though, I have to test out...