-- Model now have flag overrideProperties - if set to false, Attraction's Properties are returned on Attraction.getProperty() call - if set to true, Model's Properties are returned on Attraction.getProperty() call. Default is always set to overrideProperties=false;
NOTE: this allows for using Model as kind of a broader specification of the Attraction, eg. "Steel Coaster" or "Spinning Coaster" or "Boat Ride" independened from CreditType, Category and Manufacturer.
This is maybe complicating things a little, bit is way better than an empty Model on the Attraction imo. We will see how this turns out...